Chunking doesn't work like I thought it would!
Chunking reads the entire file it's given, it won't move to as keyframe. This update uses seek and frames to move through the file and then chunk with a 100 frame buffer on either side to grease the encoder. I believe it could be coded better for sure. 

What we want is a see/frames window that moves through the video with chunk processing within it. I start at -100 so chunk starts at frame 0. It's completely possible 100 frames may not be a good number plus or minus. As you get deeper into the video seek takes longer at job start but it's normal.