I think I got it!
Good grief why was this so hard for my pea brain? I think this works. The initial -100 frames is cancelled by the chunk start at 100 to get a real start at frame 0. The rest proceeds right on through. I've set it to end an extra frame or so after EOF which I think will be fine. Only question I've got is I end on a frame and then start on that same frame for the next section - I think this is wrong and it should start at the next frame in sequence.. That shouldn't be hard to fix but I'm excited to get it this far. I'm going to go through a full length movie and see if the audio synchs after mux.